+++ dashboard/models_executive.py
@@ -1,5 +1,6 @@
 # Dashboard Executivo - Modelos para KPIs Avançados
 from django.db import models
+from django.db.models import Q
 from django.contrib.auth.models import User
 from django.utils import timezone
 from datetime import datetime, timedelta
@@ -21,8 +22,8 @@ class ExecutiveDashboardKPI(models.Model):
     
     name = models.CharField(max_length=100, verbose_name="Nome do KPI")
     kpi_type = models.CharField(max_length=50, choices=KPI_TYPES, verbose_name="Tipo de KPI")
-    target_value = models.FloatField(verbose_name="Meta")
-    current_value = models.FloatField(default=0, verbose_name="Valor Atual")
+    target_value = models.DecimalField(max_digits=12, decimal_places=2, verbose_name="Meta")
+    current_value = models.DecimalField(max_digits=12, decimal_places=2, default=0, verbose_name="Valor Atual")
     department = models.CharField(max_length=50, blank=True, verbose_name="Departamento")
     is_active = models.BooleanField(default=True, verbose_name="Ativo")
     created_at = models.DateTimeField(auto_now_add=True)
@@ -128,7 +129,7 @@ class AlertaKPI(models.Model):
     alert_type = models.CharField(max_length=20, choices=ALERT_TYPES, verbose_name="Tipo de Alerta")
     severity = models.CharField(max_length=20, choices=SEVERITY_LEVELS, verbose_name="Severidade")
     message = models.TextField(verbose_name="Mensagem")
-    threshold_value = models.FloatField(null=True, blank=True, verbose_name="Valor Limite")
+    threshold_value = models.DecimalField(max_digits=12, decimal_places=2, null=True, blank=True, verbose_name="Valor Limite")
     is_resolved = models.BooleanField(default=False, verbose_name="Resolvido")
     created_at = models.DateTimeField(auto_now_add=True)
     resolved_at = models.DateTimeField(null=True, blank=True, verbose_name="Resolvido em")
